// src/app/@theme/components/Model/CurrentDevicePositionResponse.ts
  export interface CurrentDevicePositionResponse {
    status: boolean
    message: string
    data: CurrentDevicePositionResponse
  }
  
  export interface CurrentDevicePositionResponse {
    la: string
    lo: string
    c: string
    s: string
    dt: string
    state: string
    district: string
    l: string
    f: string
    t: string
    b: string
    vms :string
    vmsc :string
  }
  
  